Misc fixes for CentOS 6 & 8
CentOS 6 is EOL but we still need to provide builds for a little bit longer.
Upstream has moved everything to the vault repos so this updates the image to
use the vault repo instead.

In addition. CentOS 8.3 changed the names of the default OS repo files slightly,
so this accounts for those changes.

Signed-off-by: Lance Albertson <lance@osuosl.org>